[test] custom elasticsearch config for sourcemaps (#3349)
Add system tests for `apm-server.rum.source_mapping.elasticsearch.*`
configuration, including username+password and api_key settings.

closes #3166
