+++ tests/system/test_integration.py
@@ -279,262 +279,6 @@ class OverrideIndicesFailureIntegrationTest(ProcStartupFailureTest):
 
 
 @integration_test
-class SourcemappingIntegrationTest(ClientSideElasticTest):
-    def test_backend_error(self):
-        # ensure source mapping is not applied to backend events
-        # load event for which a sourcemap would be applied when sent to rum endpoint,
-        # and send against backend endpoint.
-
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(
-            file_name='bundle.js.map', bundle_filepath=path)
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.backend_intake_url,
-                                     'error',
-                                     1)
-        self.assert_no_logged_warnings()
-        self.check_backend_error_sourcemap(index_error)
-
-    def test_duplicated_sourcemap_warning(self):
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-
-        self.upload_sourcemap(file_name='bundle.js.map', bundle_filepath=path)
-        self.wait_for_sourcemaps()
-
-        self.upload_sourcemap(file_name='bundle.js.map', bundle_filepath=path)
-        self.wait_for_sourcemaps(2)
-        assert self.log_contains(
-            "Overriding sourcemap"), "A log should be written when a sourcemap is overwritten"
-
-        self.upload_sourcemap(file_name='bundle.js.map', bundle_filepath=path)
-        self.wait_for_sourcemaps(3)
-        assert self.log_contains("2 sourcemaps found for service"), \
-            "the 3rd fetch should query ES and find that there are 2 sourcemaps with the same caching key"
-
-        self.assert_no_logged_warnings(
-            ["WARN.*Overriding sourcemap", "WARN.*2 sourcemaps found for service"])
-
-    def test_rum_error(self):
-        # use an uncleaned path to test that path is cleaned in upload
-        path = 'http://localhost:8000/test/e2e/../e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(file_name='bundle.js.map', bundle_filepath=path)
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.assert_no_logged_warnings()
-        self.check_rum_error_sourcemap(True)
-
-    def test_backend_span(self):
-        # ensure source mapping is not applied to backend events
-        # load event for which a sourcemap would be applied when sent to rum endpoint,
-        # and send against backend endpoint.
-
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(file_name='bundle.js.map',
-                                  bundle_filepath=path,
-                                  service_version='1.0.0')
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-
-        self.load_docs_with_template(self.get_transaction_payload_path(),
-                                     self.backend_intake_url,
-                                     'transaction',
-                                     2)
-        self.assert_no_logged_warnings()
-        self.check_backend_span_sourcemap()
-
-    def test_rum_transaction(self):
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(file_name='bundle.js.map',
-                                  bundle_filepath=path,
-                                  service_version='1.0.0')
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-
-        self.load_docs_with_template(self.get_transaction_payload_path(),
-                                     self.intake_url,
-                                     'transaction',
-                                     2)
-        self.assert_no_logged_warnings()
-        self.check_rum_transaction_sourcemap(True)
-
-    def test_rum_transaction_different_subdomain(self):
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(file_name='bundle.js.map',
-                                  bundle_filepath=path,
-                                  service_version='1.0.0')
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-
-        self.load_docs_with_template(self.get_payload_path('transactions_spans_rum_2.ndjson'),
-                                     self.intake_url,
-                                     'transaction',
-                                     2)
-        self.assert_no_logged_warnings()
-        self.check_rum_transaction_sourcemap(True)
-
-    def test_no_sourcemap(self):
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.check_rum_error_sourcemap(
-            False, expected_err="No Sourcemap available")
-
-    def test_no_matching_sourcemap(self):
-        r = self.upload_sourcemap('bundle_no_mapping.js.map')
-        self.assert_no_logged_warnings()
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-        self.test_no_sourcemap()
-
-    def test_fetch_latest_of_multiple_sourcemaps(self):
-        # upload sourcemap file that finds no matchings
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(
-            file_name='bundle_no_mapping.js.map', bundle_filepath=path)
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.check_rum_error_sourcemap(
-            False, expected_err="No Sourcemap found for")
-
-        # remove existing document
-        self.es.delete_by_query(index=index_error, body={"query": {"term": {"processor.name": 'error'}}})
-        wait_until(lambda: (self.es.count(index=index_error)['count'] == 0))
-
-        # upload second sourcemap file with same key,
-        # that actually leads to proper matchings
-        # this also tests that the cache gets invalidated,
-        # as otherwise the former sourcemap would be taken from the cache.
-        r = self.upload_sourcemap(
-            file_name='bundle.js.map', bundle_filepath=path)
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps(expected_ct=2)
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.check_rum_error_sourcemap(True, count=1)
-
-    def test_sourcemap_mapping_cache_usage(self):
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(
-            file_name='bundle.js.map', bundle_filepath=path)
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-
-        # insert document, which also leads to caching the sourcemap
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.assert_no_logged_warnings()
-
-        # delete sourcemap from ES
-        # fetching from ES would lead to an error afterwards
-        self.es.indices.delete(index=index_smap, ignore=[400, 404])
-        self.es.indices.delete(index="{}-000001".format(index_error), ignore=[400, 404])
-
-        # insert document,
-        # fetching sourcemap without errors, so it must be fetched from cache
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.assert_no_logged_warnings()
-        self.check_rum_error_sourcemap(True)
-
-    def test_rum_error_changed_index(self):
-        # use an uncleaned path to test that path is cleaned in upload
-        path = 'http://localhost:8000/test/e2e/../e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(
-            file_name='bundle.js.map', bundle_filepath=path)
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.assert_no_logged_warnings()
-        self.check_rum_error_sourcemap(True)
-
-
-@integration_test
-class SourcemappingCacheIntegrationTest(ClientSideElasticTest):
-    config_overrides = {"smap_cache_expiration": "1"}
-
-    def test_sourcemap_cache_expiration(self):
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(
-            file_name='bundle.js.map', bundle_filepath=path)
-        assert r.status_code == 202, r.status_code
-        self.wait_for_sourcemaps()
-
-        # insert document, which also leads to caching the sourcemap
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.assert_no_logged_warnings()
-
-        # delete sourcemap and error event from ES
-        self.es.indices.delete(index=self.ilm_index(index_error))
-        # fetching from ES will lead to an error afterwards
-        self.es.indices.delete(index=index_smap, ignore=[400, 404])
-        wait_until(lambda: not self.es.indices.exists(index_smap))
-        # ensure smap is not in cache any more
-        time.sleep(1)
-
-        # after cache expiration no sourcemap should be found any more
-        self.load_docs_with_template(self.get_error_payload_path(),
-                                     self.intake_url,
-                                     'error',
-                                     1)
-        self.check_rum_error_sourcemap(False, expected_err="No Sourcemap available")
-
-
-@integration_test
-class SourcemappingDisabledIntegrationTest(ClientSideElasticTest):
-    config_overrides = {
-        "rum_sourcemapping_disabled": True,
-    }
-
-    def test_rum_transaction(self):
-        path = 'http://localhost:8000/test/e2e/general-usecase/bundle.js.map'
-        r = self.upload_sourcemap(file_name='bundle.js.map',
-                                  bundle_filepath=path,
-                                  service_version='1.0.0')
-        assert r.status_code == 403, r.status_code
-
-        self.load_docs_with_template(self.get_transaction_payload_path(),
-                                     self.intake_url,
-                                     'transaction',
-                                     2)
-        rs = self.es.search(index=index_span, params={"rest_total_hits_as_int": "true"})
-        assert rs['hits']['total'] == 1, "found {} documents, expected {}".format(
-            rs['hits']['total'], 1)
-        frames_checked = 0
-        for doc in rs['hits']['hits']:
-            span = doc["_source"]["span"]
-            for frame in span["stacktrace"]:
-                frames_checked += 1
-                assert "sourcemap" not in frame, frame
-        assert frames_checked > 0, "no frames checked"
-
-
-@integration_test
 class ExpvarDisabledIntegrationTest(ExpvarBaseTest):
     config_overrides = {"expvar_enabled": "false"}
